+++ src/App.js
@@ -15,7 +15,6 @@ import theme from './theme';
 import useLocalStorage from './use-local-storage';
 
 import About from './ui/about';
-import Impressum from './ui/impressum';
 import NotFound from './ui/not-found';
 import OpencastHeader from './ui/opencast-header';
 import Studio from './ui/studio';
@@ -84,7 +83,6 @@ function App(props) {
             handleOpenUploadSettings={handleOpenUploadSettings}
           />
           <About path="/about" />
-          <Impressum path="/impressum" />
           <NotFound default />
         </Router>
       </Flex>
